convert geometry types from float to int

+++ src/main/java/net/sf/jsi/Point.java
@@ -25,7 +25,7 @@ public class Point {
   /**
    * The (x, y) coordinates of the point.
    */
-  public float x, y;
+  public int x, y;
 
   /**
    * Constructor.
@@ -33,7 +33,7 @@ public class Point {
    * @param x The x coordinate of the point
    * @param y The y coordinate of the point
    */
-  public Point(float x, float y) {
+  public Point(int x, int y) {
     this.x = x;
     this.y = y;
   }
